requirements.txt
Flask
python-dotenv
together
PyMuPDF